pyproject.toml
[project]
name = "modbus-monitor"
version = "0.1.0"
description = "Modbus Monitor with Vue 3 frontend and FastAPI backend"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README_NEW.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pymodbus>=3.0.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.uv]
dev-dependencies = []